@@ -1,27 +1,25 @@
 {
-  "name": "homepage",
-  "version": "0.1.0",
+  "name": "waku-project",
+  "version": "0.0.0",
+  "type": "module",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "waku dev",
+    "build": "waku build",
+    "start": "waku start --port 8082"
   },
   "dependencies": {
-    "autoprefixer": "^10.4.8",
-    "next": "12.2.5",
-    "postcss": "^8.4.16",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "^3.1.8"
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-server-dom-webpack": "19.0.0",
+    "waku": "0.21.24"
   },
   "devDependencies": {
-    "@types/node": "18.7.13",
-    "@types/react": "18.0.17",
-    "@types/react-dom": "18.0.6",
-    "eslint": "8.23.0",
-    "eslint-config-next": "12.2.5",
-    "typescript": "4.8.2"
+    "@tailwindcss/postcss": "4.0.15",
+    "@types/react": "19.0.12",
+    "@types/react-dom": "19.0.4",
+    "postcss": "8.5.3",
+    "tailwindcss": "4.0.15",
+    "typescript": "5.8.2"
   }
 }
